refactor: dto 네이밍 변경 및 패키지 변경
Dto의 역할보다는 단순히 도메인 값 객체데 가까우므로 네이밍을 변경함
- dto패키지에 있지만 domain 역할을 하고있는 클래스를 도메인 패키지로 옮김
