+++ src/main/java/lotto/domain/MatchResult.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public record MatchResult(int matchCount, boolean isBonusMatched) {
+}
