+++ src/main/java/lotto/domain/Profit.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.domain;
 
 public record Profit(double rate, boolean isProfit) {
