--- src/main/java/lotto/dto/MatchCountDto.java
@@ -1,4 +1,0 @@
-package lotto.dto;
-
-public record MatchCountDto(int matchCount, boolean isBonusMatched) {
-}
