+++ src/test/java/lotto/domain/LottoStatisticsTest.java
@@ -3,7 +3,6 @@ package lotto.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import lotto.dto.Profit;
 import lotto.dto.WinningInform;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
